Update task_manager_v2.c
We have modified our UI to notify the user when a mobile is in a state called Connecting... meaning that the user is currently in retransmission. This state is often temporary since the underlying SA reestablishes itself after roaming.

We noticed that the ALERT_RETRANSMIT_SEND_CLEARED is not always sent. We added a condition which handles the situation.

The clearest example of when it is not cleared is :

- Several retransmission occurs due to being unable to reach a vpn gateway 
- Mobile IKE is defered and thus this->active_tasks is not empty and is carrying a mobike deferred task
- Connection resumes due to a change in connectivity i.e. cell or wifi is reenabled
- At which point the this->initiating.retransmitted is set to 0 by mobike.
- This thus creates a situation where the ALERT_RETRANSMIT_SEND_CLEARED is never sent. Thus, the UI will stay in connecting state.

In the end, adding a flag to let the task v2 know that a send has been sent before enables clearing the message in the UI.

The current code of strongswan does not make use of such a state in the UI but we thought to share the change since if it becomes a necessity in the future it will properly reprensent the state of the SA in the UI.